WIP Data driven guns - Separated all gun data from it's item
