+++ crafting-dead-core/src/main/java/com/craftingdead/core/world/item/gun/GunLike.java
@@ -1,0 +1,6 @@
+package com.craftingdead.core.world.item.gun;
+
+public interface GunLike {
+
+  GunType asGun();
+}
