+++ crafting-dead-core/src/main/java/com/craftingdead/core/world/item/gun/GunType.java
@@ -1,65 +1,38 @@
-/*
- * Crafting Dead
- * Copyright (C) 2022  NexusNode LTD
- *
- * This Non-Commercial Software License Agreement (the "Agreement") is made between
- * you (the "Licensee") and NEXUSNODE (BRAD HUNTER). (the "Licensor").
- * By installing or otherwise using Crafting Dead (the "Software"), you agree to be
- * bound by the terms and conditions of this Agreement as may be revised from time
- * to time at Licensor's sole discretion.
- *
- * If you do not agree to the terms and conditions of this Agreement do not download,
- * copy, reproduce or otherwise use any of the source code available online at any time.
- *
- * https://github.com/nexusnode/crafting-dead/blob/1.18.x/LICENSE.txt
- *
- * https://craftingdead.net/terms.php
- */
-
 package com.craftingdead.core.world.item.gun;
 
+import com.craftingdead.core.CraftingDead;
+import com.craftingdead.core.capability.CapabilityUtil;
+import com.craftingdead.core.client.animation.Animation;
+import com.craftingdead.core.util.FunctionalUtil;
+import com.craftingdead.core.world.item.GunItem;
+import com.craftingdead.core.world.item.combatslot.CombatSlot;
+import com.craftingdead.core.world.item.combatslot.CombatSlotProvider;
+import com.craftingdead.core.world.item.gun.ammoprovider.AmmoProvider;
+import com.craftingdead.core.world.item.gun.ammoprovider.MagazineAmmoProvider;
+import com.craftingdead.core.world.item.gun.attachment.Attachment;
 import java.util.EnumMap;
 import java.util.EnumSet;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.IntFunction;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import javax.annotation.Nullable;
-import com.craftingdead.core.CraftingDead;
-import com.craftingdead.core.client.animation.Animation;
-import com.craftingdead.core.util.FunctionalUtil;
-import com.craftingdead.core.world.item.ModItems;
-import com.craftingdead.core.world.item.RegisterGunColor;
-import com.craftingdead.core.world.item.combatslot.CombatSlot;
-import com.craftingdead.core.world.item.gun.ammoprovider.AmmoProvider;
-import com.craftingdead.core.world.item.gun.ammoprovider.MagazineAmmoProvider;
-import com.craftingdead.core.world.item.gun.attachment.Attachment;
-import com.craftingdead.core.world.item.gun.magazine.Magazine;
-import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.TextComponent;
-import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ProjectileWeaponItem;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.Enchantments;
-import net.minecraft.world.level.Level;
+import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.common.capabilities.ICapabilityProvider;
+import net.minecraftforge.registries.ForgeRegistryEntry;
+import org.jetbrains.annotations.NotNull;
 
-@RegisterGunColor
-public abstract class GunItem extends ProjectileWeaponItem {
+public class GunType extends ForgeRegistryEntry<GunType> implements ItemLike {
+
+  private final Supplier<? extends GunItem> item;
 
   /**
    * Time between shots in milliseconds.
@@ -121,7 +94,7 @@ public abstract class GunItem extends ProjectileWeaponItem {
    */
   private final Supplier<SoundEvent> reloadSound;
 
-  private final Map<GunAnimationEvent, Function<GunItem, Animation>> animations;
+  private final Map<GunAnimationEvent, Function<GunType, Animation>> animations;
 
   /**
    * A set of magazines that are supported by this gun.
@@ -165,8 +138,9 @@ public abstract class GunItem extends ProjectileWeaponItem {
 
   private final CombatSlot combatSlot;
 
-  protected GunItem(Builder<?> builder) {
-    super(builder.properties);
+  //TODO: move to properties after the split is done
+  protected GunType(Builder<?> builder) {
+    this.item = builder.item;
     this.fireDelayMs = builder.fireDelayMs;
     this.damage = builder.damage;
     this.reloadDurationTicks = builder.reloadDurationTicks;
@@ -247,7 +221,7 @@ public abstract class GunItem extends ProjectileWeaponItem {
     return Optional.ofNullable(this.reloadSound).map(Supplier::get);
   }
 
-  public Map<GunAnimationEvent, Function<GunItem, Animation>> getAnimations() {
+  public Map<GunAnimationEvent, Function<GunType, Animation>> getAnimations() {
     return this.animations;
   }
 
@@ -292,123 +266,30 @@ public abstract class GunItem extends ProjectileWeaponItem {
     return this.combatSlot;
   }
 
-  @Override
-  public Predicate<ItemStack> getAllSupportedProjectiles() {
-    return itemStack -> this.getAcceptedMagazines()
-        .stream()
-        .anyMatch(itemStack.getItem()::equals);
+  public ICapabilityProvider initCapabilities(ItemStack itemStack, CompoundTag nbt) {
+    return CapabilityUtil.serializableProvider(
+        () -> TypedGun.create(this.getClientFactory(), itemStack, this),
+        Gun.CAPABILITY, CombatSlotProvider.CAPABILITY);
+  }
+
+  public <T extends TypedGun<?>> Function<T, TypedGunClient<? super T>> getClientFactory() {
+    return TypedGunClient::new;
   }
 
   @Override
-  public abstract ICapabilityProvider initCapabilities(ItemStack itemStack,
-      @Nullable CompoundTag nbt);
-
-  @Override
-  public boolean onEntitySwing(ItemStack stack, LivingEntity entity) {
-    return true;
+  public @NotNull Item asItem() {
+    return this.item.get();
   }
 
-  @Override
-  public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack,
-      boolean slotChanged) {
-    return oldStack.getItem() != newStack.getItem();
-  }
-
-  @Override
-  public void appendHoverText(ItemStack itemStack, Level world,
-      List<Component> lines, TooltipFlag tooltipFlag) {
-    super.appendHoverText(itemStack, world, lines, tooltipFlag);
-
-    itemStack.getCapability(Gun.CAPABILITY).ifPresent(gun -> {
-      Component ammoCount =
-          new TextComponent(String.valueOf(gun.getAmmoProvider().getMagazine()
-              .map(Magazine::getSize)
-              .orElse(0))).withStyle(ChatFormatting.RED);
-      Component damageText =
-          new TextComponent(String.valueOf(this.getDamage()))
-              .withStyle(ChatFormatting.RED);
-      Component headshotDamageText = new TextComponent(
-          String.valueOf((int) (this.getDamage() * CraftingDead.serverConfig.headshotBonusDamage.get())))
-              .withStyle(ChatFormatting.RED);
-      Component accuracyText =
-          new TextComponent((int) (this.getAccuracyPct() * 100D) + "%")
-              .withStyle(ChatFormatting.RED);
-      Component rpmText =
-          new TextComponent(String.valueOf(this.getFireRateRPM()))
-              .withStyle(ChatFormatting.RED);
-      Component rangeText =
-          new TextComponent(this.getRange() + " blocks")
-              .withStyle(ChatFormatting.RED);
-
-      lines.add(new TranslatableComponent("gun.ammo_amount")
-          .withStyle(ChatFormatting.GRAY)
-          .append(ammoCount));
-      lines.add(new TranslatableComponent("gun.damage")
-          .withStyle(ChatFormatting.GRAY)
-          .append(damageText));
-      lines.add(new TranslatableComponent("gun.headshot_damage")
-          .withStyle(ChatFormatting.GRAY)
-          .append(headshotDamageText));
-
-      if (this.getRoundsPerShot() > 1) {
-        Component pelletsText =
-            new TextComponent(String.valueOf(this.getRoundsPerShot()))
-                .withStyle(ChatFormatting.RED);
-
-        lines.add(new TranslatableComponent("gun.pellets_shot")
-            .withStyle(ChatFormatting.GRAY)
-            .append(pelletsText));
-      }
-
-      for (Attachment attachment : gun.getAttachments()) {
-        Component attachmentName = attachment.getDescription()
-            .plainCopy()
-            .withStyle(ChatFormatting.RED);
-        lines.add(new TranslatableComponent("gun.attachment")
-            .withStyle(ChatFormatting.GRAY)
-            .append(attachmentName));
-      }
-
-      lines.add(new TranslatableComponent("gun.rpm")
-          .withStyle(ChatFormatting.GRAY)
-          .append(rpmText));
-      lines.add(new TranslatableComponent("gun.accuracy")
-          .withStyle(ChatFormatting.GRAY)
-          .append(accuracyText));
-      lines.add(new TranslatableComponent("gun.range")
-          .withStyle(ChatFormatting.GRAY)
-          .append(rangeText));
-    });
-  }
-
-  @Override
-  public boolean isEnchantable(ItemStack stack) {
-    return true;
-  }
-
-  @Override
-  public boolean canApplyAtEnchantingTable(ItemStack stack, Enchantment enchantment) {
-    return enchantment == Enchantments.FLAMING_ARROWS
-        || enchantment == Enchantments.POWER_ARROWS
-        || enchantment == Enchantments.UNBREAKING
-        || super.canApplyAtEnchantingTable(stack, enchantment);
-  }
-
-  @Override
-  public int getEnchantmentValue() {
-    return 1;
-  }
-
-  @Override
-  public int getDefaultProjectileRange() {
-    return 0;
+  public static Builder<?> builder() {
+    return new Builder<>(GunType::new);
   }
 
   public static class Builder<SELF extends Builder<SELF>> {
 
-    private final Function<SELF, GunItem> factory;
+    private final Function<SELF, GunType> factory;
 
-    private final Properties properties = new Properties().stacksTo(1).tab(ModItems.COMBAT_TAB);
+    private Supplier<? extends GunItem> item;
 
     private int fireDelayMs;
 
@@ -436,7 +317,7 @@ public abstract class GunItem extends ProjectileWeaponItem {
 
     private Supplier<SoundEvent> reloadSound = () -> null;
 
-    private final Map<GunAnimationEvent, Function<GunItem, Animation>> animations =
+    private final Map<GunAnimationEvent, Function<GunType, Animation>> animations =
         new EnumMap<>(GunAnimationEvent.class);
 
     private final Set<Supplier<? extends Item>> acceptedMagazines = new HashSet<>();
@@ -456,14 +337,15 @@ public abstract class GunItem extends ProjectileWeaponItem {
 
     private CombatSlot combatSlot = CombatSlot.PRIMARY;
 
-    public Builder(Function<SELF, GunItem> factory) {
+    public Builder(Function<SELF, GunType> factory) {
       this.factory = factory;
     }
 
-    public SELF properties(Consumer<Properties> consumer) {
-      consumer.accept(this.properties);
+    public SELF setItem(Supplier<? extends GunItem> item) {
+      this.item = item  ;
       return this.self();
     }
+
 
     public SELF setFireDelayMs(int fireDelayMs) {
       this.fireDelayMs = fireDelayMs;
@@ -540,7 +422,7 @@ public abstract class GunItem extends ProjectileWeaponItem {
     }
 
     public SELF putAnimation(GunAnimationEvent event,
-        Function<GunItem, Animation> animation) {
+        Function<GunType, Animation> animation) {
       this.animations.put(event, animation);
       return this.self();
     }
@@ -590,7 +472,7 @@ public abstract class GunItem extends ProjectileWeaponItem {
       return this.self();
     }
 
-    public GunItem build() {
+    public GunType build() {
       return this.factory.apply(this.self());
     }
